LPS-110416 Decouple, TableReferenceDefinitionManager should only care about data collection, and let CTClosureFactory impl cares about how to generate closure with data
