+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/reference/closure/CTClosureFactoryImpl.java
@@ -1,0 +1,69 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.change.tracking.internal.reference.closure;
+
+import com.liferay.change.tracking.internal.reference.SchemaContext;
+import com.liferay.change.tracking.internal.reference.TableReferenceDefinitionManager;
+import com.liferay.change.tracking.internal.reference.closure.CTClosureImpl;
+import com.liferay.change.tracking.internal.reference.closure.Node;
+import com.liferay.change.tracking.model.CTEntry;
+import com.liferay.change.tracking.reference.closure.CTClosure;
+import com.liferay.change.tracking.reference.closure.CTClosureFactory;
+import com.liferay.change.tracking.service.CTEntryLocalService;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Tina Tian
+ */
+@Component(immediate = true, service = CTClosureFactory.class)
+public class CTClosureFactoryImpl implements CTClosureFactory {
+
+	@Override
+	public CTClosure create(long ctCollectionId) {
+		Map<Long, List<Long>> map = new HashMap<>();
+
+		for (CTEntry ctEntry :
+				_ctEntryLocalService.getCTCollectionCTEntries(ctCollectionId)) {
+
+			List<Long> primaryKeys = map.computeIfAbsent(
+				ctEntry.getModelClassNameId(), key -> new ArrayList<>());
+
+			primaryKeys.add(ctEntry.getModelClassPK());
+		}
+
+		SchemaContext schemaContext =
+			_tableReferenceDefinitionManager.getSchemaContext();
+
+		Map<Node, Collection<Node>> closureMap = schemaContext.buildClosureMap(
+			ctCollectionId, map);
+
+		return new CTClosureImpl(ctCollectionId, closureMap);
+	}
+
+	@Reference
+	private CTEntryLocalService _ctEntryLocalService;
+
+	@Reference
+	private TableReferenceDefinitionManager _tableReferenceDefinitionManager;
+
+}
